FF-33 #comment Minor edits - fix re: React Invariant Error (Date -> UTC timezone)
