+++ .ebextensions/01_packages.config
@@ -12,7 +12,7 @@ packages:
 
 container_commands:
   01_nodejs_install:
-    command: curl --silent --location https://rpm.nodesource.com/setup_7.x | sudo bash - && yum install nodejs -y
+    command: curl --silent --location https://rpm.nodesource.com/setup_7.x | sudo bash - && yum install nodejs -y && node --version
   02_setup_apache:
     command: cp .ebextensions/encoded-apache.conf /etc/httpd/conf.d/encoded-apache.conf
   03_bootstrap:
