+++ buildout.cfg
@@ -172,7 +172,7 @@ cmds = NODE_PATH="" npm_config_cache="" npm install
 recipe = collective.recipe.cmd
 on_install = true
 on_update = true
-cmds = NODE_PATH="" npm_config_cache="" npm update
+cmds = NODE_PATH="" npm_config_cache="" npm update --dependencies
 
 [compile-js]
 recipe = collective.recipe.cmd
